Move UML related drop() code to UML package
